+++ src/Typin/Typin.Core/Console/IO/IStandardOutput.cs
@@ -1,0 +1,13 @@
+﻿namespace Typin.Console.IO
+{
+    /// <summary>
+    /// Abstraction for interacting with the standard output stream (stdout).
+    /// </summary>
+    public interface IStandardOutput
+    {
+        /// <summary>
+        /// Output stream (stdout) writer.
+        /// </summary>
+        StandardStreamWriter Output { get; }
+    }
+}