+++ src/Typin/Typin.Tests/DirectivesTests.cs
@@ -69,6 +69,49 @@
         }
 
         [Fact]
+        public async Task Preview_directive_can_be_specified_before_debug_directive()
+        {
+            // Arrange
+            var builder = new CliApplicationBuilder()
+                .AddCommand<NamedCommand>()
+                .UseInteractiveMode()
+                .AddDirective<PreviewDirective>()
+                .AddDirective<DebugDirective>();
+
+            // Act
+            var (exitCode, stdOut, stdErr) = await builder.BuildAndRunTestAsync(_output,
+                new[] { "[preview]", "[debug]", "named", "param", "-abc", "--option", "foo" });
+
+            // Assert
+            exitCode.Should().Be(ExitCodes.Success);
+            stdOut.GetString().Should().NotBeNullOrWhiteSpace();
+            stdOut.GetString().Should().NotContain("Attach debugger to PID");
+            stdErr.GetString().Should().BeNullOrWhiteSpace();
+        }
+
+        [Fact]
+        public async Task Preview_directive_should_work_in_direct_mode_even_if_directives_from_other_mode_are_specified()
+        {
+            // Arrange
+            var builder = new CliApplicationBuilder()
+                .AddCommand<NamedCommand>()
+                .UseInteractiveMode()
+                .AddDirective<PreviewDirective>()
+                .AddDirective<DebugDirective>(); //TODO: add test when UseInteractiv and AddDirective<ScopeUp> are used and check if error is thrown
+            //TODO: what if unknown directive is passed after [preview]?
+
+            // Act
+            var (exitCode, stdOut, stdErr) = await builder.BuildAndRunTestAsync(_output,
+                new[] { "[preview]", "[debug]", "named", "param", "-abc", "--option", "foo" });
+
+            // Assert
+            exitCode.Should().Be(ExitCodes.Success);
+            stdOut.GetString().Should().NotBeNullOrWhiteSpace();
+            stdOut.GetString().Should().NotContain("Attach debugger to PID");
+            stdErr.GetString().Should().BeNullOrWhiteSpace();
+        }
+
+        [Fact]
         public async Task Custom_stop_directive_should_cancel_execution_after_running()
         {
             // Arrange
@@ -83,7 +126,7 @@
                 new[] { "[custom-stop]", "named", "param", "-abc", "--option", "foo" });
 
             // Assert
-            exitCode.Should().Be(ExitCodes.Success);
+            exitCode.Should().Be(CustomStopDirective.ExpectedExitCode);
             stdOut.GetString().Should().NotBeNullOrWhiteSpace();
             stdOut.GetString().Should().Be(CustomStopDirective.ExpectedOutput);
         }
@@ -149,7 +192,7 @@
                 new[] { "[custom-interactive]", "named", "param", "-abc", "--option", "foo" });
 
             // Assert
-            exitCode.Should().NotBe(0);
+            exitCode.Should().NotBe(ExitCodes.Success);
             stdOut.GetString().Should().BeNullOrWhiteSpace();
             stdOut.GetString().Should().NotContainAll(
                 "@ [custom-interactive]", "Description", "Usage", "Directives", "[custom]"