+++ src/Typin/Typin/CliApplicationBuilder.cs
@@ -7,6 +7,7 @@ namespace Typin
     using System.Text.RegularExpressions;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.DependencyInjection.Extensions;
+    using Microsoft.Extensions.Logging;
     using Typin.Console;
     using Typin.Exceptions;
     using Typin.HelpWriter;
@@ -55,7 +56,7 @@ namespace Typin
         /// </summary>
         public CliApplicationBuilder()
         {
-
+            this.AddAfterInputParseMiddlewares();
         }
 
         #region Directives
@@ -338,69 +339,6 @@ namespace Typin
         }
         #endregion
 
-        #region Configuration
-        //TODO add configuration builder on actions https://github.com/aspnet/Hosting/blob/f9d145887773e0c650e66165e0c61886153bcc0b/src/Microsoft.Extensions.Hosting/HostBuilder.cs
-
-        /// <summary>
-        /// Configures application services.
-        /// </summary>
-        public CliApplicationBuilder Configure(Action<CliApplicationBuilder> action)
-        {
-            action.Invoke(this);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Configures application services.
-        /// </summary>
-        public CliApplicationBuilder ConfigureServices(Action<IServiceCollection> action)
-        {
-            _configureServicesActions.Add(action);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Configures application using <see cref="ICliStartup"/> class instance.
-        /// </summary>
-        public CliApplicationBuilder UseStartup<T>()
-            where T : class, ICliStartup, new()
-        {
-            ICliStartup t = new T();
-            _configureServicesActions.Add(t.ConfigureServices);
-            t.Configure(this);
-
-            return this;
-        }
-
-
-        //https://github.com/aspnet/Hosting/blob/f9d145887773e0c650e66165e0c61886153bcc0b/src/Microsoft.Extensions.Hosting/HostBuilder.cs
-
-        /// <summary>
-        /// Overrides the factory used to create the service provider.
-        /// </summary>
-        public CliApplicationBuilder UseServiceProviderFactory<TContainerBuilder>(IServiceProviderFactory<TContainerBuilder> factory)
-            where TContainerBuilder : notnull
-        {
-            _serviceProviderAdapter = new ServiceFactoryAdapter<TContainerBuilder>(factory ?? throw new ArgumentNullException(nameof(factory)));
-
-            return this;
-        }
-
-        /// <summary>
-        /// Enables configuring the instantiated dependency container. This can be called multiple times and
-        /// the results will be additive.
-        /// </summary>
-        public CliApplicationBuilder ConfigureContainer<TContainerBuilder>(Action<TContainerBuilder> configureDelegate)
-        {
-            _configureContainerActions.Add(new ConfigureContainerAdapter<TContainerBuilder>(configureDelegate
-                ?? throw new ArgumentNullException(nameof(configureDelegate))));
-
-            return this;
-        }
-        #endregion
-
         #region Middleware
         /// <summary>
         /// Adds a middleware to the command execution pipeline.
@@ -414,7 +352,7 @@ namespace Typin
                 services.AddScoped(middleware);
             });
 
-            _middlewareTypes.AddFirst(middleware);
+            _middlewareTypes.AddLast(middleware);
 
             return this;
         }
@@ -477,6 +415,82 @@ namespace Typin
         }
         #endregion
 
+        #region Configuration
+        //TODO: add configuration builder on actions https://github.com/aspnet/Hosting/blob/f9d145887773e0c650e66165e0c61886153bcc0b/src/Microsoft.Extensions.Hosting/HostBuilder.cs
+
+        /// <summary>
+        /// Configures application services.
+        /// </summary>
+        public CliApplicationBuilder Configure(Action<CliApplicationBuilder> action)
+        {
+            action.Invoke(this);
+
+            return this;
+        }
+
+        /// <summary>
+        /// Configures application services.
+        /// </summary>
+        public CliApplicationBuilder ConfigureServices(Action<IServiceCollection> action)
+        {
+            _configureServicesActions.Add(action);
+
+            return this;
+        }
+
+        /// <summary>
+        /// Configures application services.
+        /// </summary>
+        public CliApplicationBuilder ConfigureLogging(Action<ILoggingBuilder> action)
+        {
+            _configureServicesActions.Add(services =>
+            {
+                services.AddLogging(action);
+            });
+
+            return this;
+        }
+
+        /// <summary>
+        /// Configures application using <see cref="ICliStartup"/> class instance.
+        /// </summary>
+        public CliApplicationBuilder UseStartup<T>()
+            where T : class, ICliStartup, new()
+        {
+            ICliStartup t = new T();
+            _configureServicesActions.Add(t.ConfigureServices);
+            t.Configure(this);
+
+            return this;
+        }
+
+
+        //https://github.com/aspnet/Hosting/blob/f9d145887773e0c650e66165e0c61886153bcc0b/src/Microsoft.Extensions.Hosting/HostBuilder.cs
+
+        /// <summary>
+        /// Overrides the factory used to create the service provider.
+        /// </summary>
+        public CliApplicationBuilder UseServiceProviderFactory<TContainerBuilder>(IServiceProviderFactory<TContainerBuilder> factory)
+            where TContainerBuilder : notnull
+        {
+            _serviceProviderAdapter = new ServiceFactoryAdapter<TContainerBuilder>(factory ?? throw new ArgumentNullException(nameof(factory)));
+
+            return this;
+        }
+
+        /// <summary>
+        /// Enables configuring the instantiated dependency container. This can be called multiple times and
+        /// the results will be additive.
+        /// </summary>
+        public CliApplicationBuilder ConfigureContainer<TContainerBuilder>(Action<TContainerBuilder> configureDelegate)
+        {
+            _configureContainerActions.Add(new ConfigureContainerAdapter<TContainerBuilder>(configureDelegate
+                ?? throw new ArgumentNullException(nameof(configureDelegate))));
+
+            return this;
+        }
+        #endregion
+
         /// <summary>
         /// Creates an instance of <see cref="CliApplication"/> using configured parameters.
         /// Default values are used in place of parameters that were not specified.
@@ -517,7 +531,7 @@ namespace Typin
             }
 
             // Add core middlewares to the end of the pipeline
-            AddCoreMiddlewares();
+            this.AddAfterUserMiddlewares();
 
             // Create context
             var _serviceCollection = new ServiceCollection();
@@ -541,18 +555,16 @@ namespace Typin
             _serviceCollection.AddSingleton<ICliCommandExecutor, CliCommandExecutor>();
             _serviceCollection.AddSingleton<ICliApplicationLifetime, CliApplicationLifetime>();
 
+            _serviceCollection.AddLogging(cfg =>
+            {
+                cfg.ClearProviders();
+                cfg.AddDebug();
+                cfg.SetMinimumLevel(LogLevel.Information);
+            });
+
             IServiceProvider serviceProvider = CreateServiceProvider(_serviceCollection);
 
             return new CliApplication(serviceProvider, cliContextFactory);
-        }
-
-        private void AddCoreMiddlewares()
-        {
-            UseMiddleware<ResolveCommandSchema>();
-            UseMiddleware<HandleVersionOption>();
-            UseMiddleware<ResolveCommandInstance>();
-            UseMiddleware<HandleHelpOption>();
-            UseMiddleware<ExecuteCommand>();
         }
 
         private IServiceProvider CreateServiceProvider(ServiceCollection services)