+++ src/Typin/Typin/Console/VirtualConsole.cs
@@ -4,6 +4,7 @@
     using System.Diagnostics.CodeAnalysis;
     using System.IO;
     using System.Threading;
+    using Typin.Console.IO;
     using Typin.Extensions;
     using Typin.Utilities;
 
@@ -18,22 +19,13 @@
         private bool disposedValue;
 
         /// <inheritdoc />
-        public StreamReader Input { get; }
+        public StandardStreamReader Input { get; }
 
         /// <inheritdoc />
-        public bool IsInputRedirected { get; }
+        public StandardStreamWriter Output { get; }
 
         /// <inheritdoc />
-        public StreamWriter Output { get; }
-
-        /// <inheritdoc />
-        public bool IsOutputRedirected { get; }
-
-        /// <inheritdoc />
-        public StreamWriter Error { get; }
-
-        /// <inheritdoc />
-        public bool IsErrorRedirected { get; }
+        public StandardStreamWriter Error { get; }
 
         /// <inheritdoc />
         public ConsoleColor ForegroundColor { get; set; } = ConsoleColor.Gray;
@@ -46,37 +38,16 @@
         /// Initializes an instance of <see cref="VirtualConsole"/>.
         /// Use named parameters to specify the streams you want to override.
         /// </summary>
-        public VirtualConsole(StreamReader? input = null, bool isInputRedirected = true,
-                              StreamWriter? output = null, bool isOutputRedirected = true,
-                              StreamWriter? error = null, bool isErrorRedirected = true,
-                              CancellationToken cancellationToken = default)
-        {
-            Input = input ?? StreamReader.Null;
-            IsInputRedirected = isInputRedirected;
-
-            Output = output ?? StreamWriter.Null;
-            IsOutputRedirected = isOutputRedirected;
-
-            Error = error ?? StreamWriter.Null;
-            IsErrorRedirected = isErrorRedirected;
-
-            _cancellationToken = cancellationToken;
-        }
-
-        /// <summary>
-        /// Initializes an instance of <see cref="VirtualConsole"/>.
-        /// Use named parameters to specify the streams you want to override.
-        /// </summary>
         public VirtualConsole(Stream? input = null, bool isInputRedirected = true,
                               Stream? output = null, bool isOutputRedirected = true,
                               Stream? error = null, bool isErrorRedirected = true,
                               CancellationToken cancellationToken = default)
-            : this(WrapInput(input), isInputRedirected,
-                   WrapOutput(output), isOutputRedirected,
-                   WrapOutput(error), isErrorRedirected,
-                   cancellationToken)
         {
+            Input = WrapInput(this, input, isInputRedirected);
+            Output = WrapOutput(this, output, isOutputRedirected);
+            Error = WrapOutput(this, error, isErrorRedirected);
 
+            _cancellationToken = cancellationToken;
         }
 
         /// <summary>
@@ -92,10 +63,10 @@
             var output = new MemoryStreamWriter(Console.OutputEncoding);
             var error = new MemoryStreamWriter(Console.OutputEncoding);
 
-            var console = new VirtualConsole(isInputRedirected: isInputRedirected,
-                                             output: output, isOutputRedirected: isOutputRedirected,
-                                             error: error, isErrorRedirected: isErrorRedirected,
-                                             cancellationToken: cancellationToken);
+            var console = new VirtualConsole(input: null, isInputRedirected,
+                                             output.Stream, isOutputRedirected,
+                                             error.Stream, isErrorRedirected,
+                                             cancellationToken);
 
             return (console, output, error);
         }
@@ -179,20 +150,20 @@
         }
 
         #region Helpers
-        private static StreamReader WrapInput(Stream? stream)
+        private static StandardStreamReader WrapInput(IConsole console, Stream? stream, bool isRedirected)
         {
             if (stream is null)
-                return StreamReader.Null;
+                return StandardStreamReader.CreateNull(console);
 
-            return new StreamReader(Stream.Synchronized(stream), Console.InputEncoding, false);
+            return new StandardStreamReader(Stream.Synchronized(stream), Console.InputEncoding, false, isRedirected, console);
         }
 
-        private static StreamWriter WrapOutput(Stream? stream)
+        private static StandardStreamWriter WrapOutput(IConsole console, Stream? stream, bool isRedirected)
         {
             if (stream is null)
-                return StreamWriter.Null;
+                return StandardStreamWriter.CreateNull(console);
 
-            return new StreamWriter(Stream.Synchronized(stream), Console.OutputEncoding)
+            return new StandardStreamWriter(Stream.Synchronized(stream), Console.OutputEncoding, isRedirected, console)
             {
                 AutoFlush = true
             };