+++ src/Typin/Typin/Directives/ScopeDirective.cs
@@ -2,9 +2,10 @@
 {
     using System.Diagnostics.CodeAnalysis;
     using System.Linq;
+    using System.Threading;
     using System.Threading.Tasks;
+    using Microsoft.Extensions.Options;
     using Typin.Attributes;
-    using Typin.Console;
     using Typin.Input;
     using Typin.Internal.Extensions;
     using Typin.Modes;
@@ -29,30 +30,36 @@
     /// </summary>
     [ExcludeFromCodeCoverage]
     [Directive(BuiltInDirectives.Scope, Description = "Sets a scope to command(s).", SupportedModes = new[] { typeof(InteractiveMode) })]
-    public sealed class ScopeDirective : IDirective
+    public sealed class ScopeDirective : IPipelinedDirective
     {
         private readonly InteractiveModeSettings _settings;
         private readonly ICliContext _cliContext;
 
-        /// <inheritdoc/>
-        public bool ContinueExecution => false;
-
         /// <summary>
         /// Initializes an instance of <see cref="ScopeDirective"/>.
         /// </summary>
-        public ScopeDirective(InteractiveModeSettings settings, ICliContext cliContext)
+        public ScopeDirective(IOptions<InteractiveModeSettings> settings, ICliContext cliContext)
         {
-            _settings = settings;
+            _settings = settings.Value;
             _cliContext = cliContext;
         }
 
         /// <inheritdoc/>
-        public ValueTask HandleAsync(IConsole console)
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
         {
-            string? name = _cliContext.Input.CommandName ?? GetFallbackCommandName();
+            return default;
+        }
+
+        /// <inheritdoc/>
+        public ValueTask HandleAsync(ICliContext context, CommandPipelineHandlerDelegate next, CancellationToken cancellationToken)
+        {
+            string? name = _cliContext.Input.CommandName ?? GetFallbackCommandName(); //TODO: fix scope directives hadnling by interactive mode
 
             if (name != null)
+            {
                 _settings.Scope = name;
+                context.ExitCode ??= ExitCodes.Success;
+            }
 
             return default;
         }