+++ src/Typin/Typin/Directives/ScopeResetDirective.cs
@@ -1,9 +1,10 @@
 ﻿namespace Typin.Directives
 {
     using System.Diagnostics.CodeAnalysis;
+    using System.Threading;
     using System.Threading.Tasks;
+    using Microsoft.Extensions.Options;
     using Typin.Attributes;
-    using Typin.Console;
     using Typin.Modes;
 
     /// <summary>
@@ -17,25 +18,29 @@
     /// </summary>
     [ExcludeFromCodeCoverage]
     [Directive(BuiltInDirectives.ScopeReset, Description = "Resets the scope to default value.", SupportedModes = new[] { typeof(InteractiveMode) })]
-    public sealed class ScopeResetDirective : IDirective
+    public sealed class ScopeResetDirective : IPipelinedDirective
     {
         private readonly InteractiveModeSettings _settings;
-
-        /// <inheritdoc/>
-        public bool ContinueExecution => false;
 
         /// <summary>
         /// Initializes an instance of <see cref="ScopeResetDirective"/>.
         /// </summary>
-        public ScopeResetDirective(InteractiveModeSettings settings)
+        public ScopeResetDirective(IOptions<InteractiveModeSettings> settings)
         {
-            _settings = settings;
+            _settings = settings.Value;
         }
 
         /// <inheritdoc/>
-        public ValueTask HandleAsync(IConsole console)
+        public ValueTask OnInitializedAsync(CancellationToken cancellationToken)
+        {
+            return default;
+        }
+
+        /// <inheritdoc/>
+        public ValueTask HandleAsync(ICliContext context, CommandPipelineHandlerDelegate next, CancellationToken cancellationToken)
         {
             _settings.Scope = string.Empty;
+            context.ExitCode ??= ExitCodes.Success;
 
             return default;
         }