+++ src/Typin/Typin/Utilities/MemoryStreamWriter.cs
@@ -2,33 +2,29 @@
 {
     using System.IO;
     using System.Text;
+    using Typin.Console;
 
     /// <summary>
-    /// Implementation of <see cref="StreamWriter"/> with a <see cref="MemoryStream"/> as a backing store.
+    /// Wrapper over a <see cref="MemoryStream"/> for <see cref="VirtualConsole"/>.
     /// </summary>
-    public class MemoryStreamWriter : StreamWriter
+    public class MemoryStreamWriter
     {
         /// <summary>
-        /// Gets the underlying stream that interfaces with a backing store.
+        /// Gets the stream that interfaces with a backing store.
         /// </summary>
-        private new MemoryStream BaseStream => (MemoryStream)base.BaseStream;
+        public MemoryStream Stream { get; } = new MemoryStream();
+
+        /// <summary>
+        /// Gets the stream that interfaces with a backing store.
+        /// </summary>
+        public Encoding Encoding { get; }
 
         /// <summary>
         /// Initializes an instance of <see cref="MemoryStreamWriter"/>.
         /// </summary>
         public MemoryStreamWriter(Encoding encoding)
-            : base(new MemoryStream(), encoding)
         {
-
-        }
-
-        /// <summary>
-        /// Initializes an instance of <see cref="MemoryStreamWriter"/>.
-        /// </summary>
-        public MemoryStreamWriter()
-            : base(new MemoryStream())
-        {
-
+            Encoding = encoding;
         }
 
         /// <summary>
@@ -36,8 +32,8 @@
         /// </summary>
         public byte[] GetBytes()
         {
-            Flush();
-            return BaseStream.ToArray();
+            Stream.Flush();
+            return Stream.ToArray();
         }
 
         /// <summary>
@@ -45,7 +41,17 @@
         /// </summary>
         public string GetString()
         {
-            return Encoding.GetString(GetBytes());
+            byte[] bytes = GetBytes();
+            return Encoding.GetString(bytes);
+        }
+
+        /// <summary>
+        /// Gets the string written to the underlying stream.
+        /// </summary>
+        public string GetString(Encoding encoding)
+        {
+            byte[] bytes = GetBytes();
+            return encoding.GetString(bytes);
         }
     }
 }