+++ CHANGELOG.md
@@ -6,9 +6,19 @@
 - Renamed `normal mode` to `direct mode`.
 - It is now possible to register multiple exception handleres to handle different exceptions in app.
 - Major API and command execution changes: a) added `ICliApplicationLifetime`, `ICliMode`, and more; b) removed `InteractiveCliApplication`.
+- Removed `HandleInteractiveDirective` and `HandleInteractiveCommands` middlewares.
 - Replaced `IsInteractiveModeOnly` with `SupportedModes`.
 - Added support for options with no name by automatic conversion of property names.
 - Added native support for .NET 5.0.
+- Added `Typin.Console.IO` namespace with `IStandardInput`, `IStandardOuput`, `IStandardError`, `IStandardRedirectableConsoleStream`, `StandardStreamReader`, `StandardStreamWriter`.
+- User middlewares are now executed after command instance creation.
+- Middleware types collection in `ApplicationConfiguration` order was reversed.
+- Merged `HandleVersionOption` and `HandleHelpOption` into one middleware named `HandleSpecialOptions`.
+- Core middleware execution order has changed: `ResolveCommandSchema` -> `ResolveCommandInstance` -> [User middlewares] -> `HandleSpecialOptions` -> `ExecuteCommand`).
+- Removed unnecessary casts to `CliContext` from `ICliContext`.
+- Removed `IDirective.ContinueExecution`, modified `IDirective`, and added `IPipelinedDirective`.
+- `CommandPipelineHandlerDelegate` now uses `ValueTask` instead of a `Task`.
+- Added experimetnal logging with `Microsoft.Extensions.Logging` (default logger is Debug).
 
 ### v2.1.1 (18-Oct-2020)
 
@@ -59,7 +69,7 @@
 - Rewritten `RootSchema` with HashSet for faster execution, esspecially in interactive mode.
 - Added tests of the command used in benchmarking to easily check if it executs correctly and won't cause banchmarking freezing.
 - Improved code readability.
-- Removed `CliApplicationBuilder.UseTypeActivator` and added Microsoft.Extensions.DependencyInjection
+- Removed `CliApplicationBuilder.UseTypeActivator` and added `Microsoft.Extensions.DependencyInjection`
 - Added support for middlewares.
 - Added `TableUtils` and `TextUtils`.
 - Added history and auto-completion in interactive mode.